chapter_001/src/main/java/ru/kambulov/condition/DummyBot.java
package ru.kambulov.condition;
/**
 * Класс будет реагировать на полученые сообщения от пользователя.
 * @author Kambulov Ivan (mailto:kia289@mail.ru)
 * @version $Id$
 * @since 0.1
 */
public class DummyBot {
    /**
     * Отвечает на вопросы.
     * "@param question Вопрос от клиента.
     * @return Ответ.
     */
    public String answer(String question) {
        String rsl = "Это ставит меня в тупик. Спросите что нибудь другое.";
        if ("Привет, Бот.".equals(question)) {
            rsl = "Привет, умник.";
        } else if ("Пока.".equals(question)) {
            rsl = "До скорой встречи!";
        }
        return rsl;
    }
}
