chapter_001/src/test/java/ru/kambulov/condition/DummyBotTest.java
package ru.kambulov.condition;

import org.junit.Test;
import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;
/**
 * Класс для тестирования методов бота.
 * @author Kambulov Ivan (mailto:kia289@mail.ru)
 * @version $Id$
 * @since 0.1
 */
public class DummyBotTest {
    @Test
    public void whenGreetBot() {
        DummyBot bot = new DummyBot();
        assertThat(bot.answer("Привет, Бот."),
        is("Привет, умник.")
        );
    }
    @Test
    public void whenByeBot() {
        DummyBot bot = new DummyBot();
        assertThat(bot.answer("Пока."),
                is("До скорой встречи!")
        );
    }
    @Test
    public void whenUnknownBot() {
        DummyBot bot = new DummyBot();
        assertThat(bot.answer("Сколько сейчас время?"),
                is("Это ставит меня в тупик. Спросите что нибудь другое.")
        );
    }
}
